Fix unnecessary integer overflow in KisTileDataStore
Casting to the wider type before multiplying will prevent
unnecessary overflow.
CID 248591

+++ libs/image/tiles3/kis_tile_data_store.cc
@@ -108,7 +108,7 @@ KisTileDataStore::MemoryStatistics KisTileDataStore::memoryStatistics()
 
     MemoryStatistics stats;
 
-    const qint64 metricCoeff = KisTileData::WIDTH * KisTileData::HEIGHT;
+    const qint64 metricCoeff = qint64(KisTileData::WIDTH) * KisTileData::HEIGHT;
 
     stats.realMemorySize = m_pooler.lastRealMemoryMetric() * metricCoeff;
     stats.historicalMemorySize = m_pooler.lastHistoricalMemoryMetric() * metricCoeff;
